requirements.txt
streamlit==1.37.1
PyYAML==6.0.2
